Recording/requirements.txt
aiohappyeyeballs==2.4.6
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
future==1.0.0
idna==3.10
iso8601==2.1.0
multidict==6.1.0
propcache==0.3.0
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
serial==0.0.97
urllib3==2.3.0
yarl==1.18.3
